requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
claude-agent-sdk==0.1.66
click==8.3.3
cryptography==46.0.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
macholib==1.16.4
mcp==1.27.0
packaging==26.1
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pyinstaller==6.20.0
pyinstaller-hooks-contrib==2026.4
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.26
referencing==0.37.0
rpds-py==0.30.0
sse-starlette==3.3.4
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
